chore(refactor): clean up code
add better names
move hanging files into utils
fix tests
fix import paths
