+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './shape'
 export * from './page'
-export * from './prepare'
-export * from './system-shapes'
+export * from './schema'
+export * from './utils/systemModels'
 export * from './utils'
